Once you enter app mode, you can never leave
In App Mode (when the program is full screen), hide the button that toggles app mode, so you can't leave it again.

This is a hacky way to make it harder for users to see the source of a program.  The purpose of this is to give us a way to give students demo programs for labs without giving them the solution to the lab.